Fix percentage based stroke-width in SVG shapes upon viewbox change
When svg elements are rendered with relative (percentage based)
stroke-width, it isn't being recalculated on viewbox change.
This poses an issue when the size of the viewbox changes such that
the size of new viewbox differs from the original viewbox. In this
scenario, the paint isn't invalidated and thus stroke is scaled
wrongly and not adjusting well with the viewport.

We have implemented a `bool ComputeHasRelativeLengths()` and used it
in the LayoutSVGShape such that it uses the stroke-width data to
determine if the current SVG element Layout has relative values and
uses the same to trigger paint recalculation upon viewbox update.

Bug: 384605094
Change-Id: I095c3d927af99e2b9ff1bf4acf086b61ed6d980a