Reland "[carousel] Add invalidation logic and tests for scroll-marker-contain"
This is a reland of commit 6b8e47152a03bde13cba00dbe76d80bc33c055df

The fix is not to traverse the whole tree to update relations when it's
known that there are no elements with scroll-marker-contain property.

Also, don't retrieve the url of <a> during AttachLayoutTree to check for
fragment presence, as it involves parsing it and affects performance.

Original change's description:
> [carousel] Add invalidation logic and tests for scroll-marker-contain
>
> Bug: 398065922
> Change-Id: I1e229d9684dadbd020b9184f175c3c527cb12788
> Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6387738
> Auto-Submit: Daniil Sakhapov <sakhapov@chromium.org>
> Commit-Queue: Daniil Sakhapov <sakhapov@chromium.org>
> Reviewed-by: Rune Lillesveen <futhark@chromium.org>
> Cr-Commit-Position: refs/heads/main@{#1438454}

Bug: 398065922
Change-Id: I90d5e6415315e3a6d2f6d12f056db3f761b38708