Expose "outbound-rtp.encodingIndex" to the web platform.
This CL adds WPT test coverage for encodingIndex and exposes it from
WebRTC.

While writing this test I discovered a bug related to RTP stats
lifetimes regarding multiple outbound-rtp being slow to show up in
getStats(). In order to decouple "testing encodingIndex in the case of
simulcast" and "RTP stats lifetimes" which is a separate issue
(chromium:406585888), we "await pollUntilMultipleOutboundRtps(pc1);"
which you shouldn't have to do but it unblocks having test coverage
for what is important.

Tests will pass pending the WebRTC roll of
https://webrtc-review.googlesource.com/c/src/+/383901.

Bug: chromium:406922375
Change-Id: I4a6a6d4f52c5c84471b5393fa5a7ed52fba9fd60